algorithms/requirements.txt
absl-py==0.11.0
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
conda==4.3.16
conda-build==2.1.5
configparser==5.0.2
cycler==0.10.0
Cython==0.29.23
cytoolz==0.11.0
deepdiff==5.2.3
docker-pycreds==0.4.0
docopt==0.6.2
enum34==1.1.10
fasteners==0.16
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
glfw==2.1.0
gym==0.10.5
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
Jinja2==3.0.0
jsonpickle==1.5.2
kiwisolver
MarkupSafe==2.0.0
matplotlib
mock==4.0.3
more-itertools==8.7.0
mpyq==0.2.5
mujoco-py==2.0.2.13
munch==2.5.0
numpy
olefile
ordered-set==4.0.2
packaging==20.9
pathlib2==2.3.5
pathtools==0.1.2
pbr==5.5.1
Pillow==7.2.0
pluggy==0.13.1
portpicker==1.3.1
probscale==0.2.4
promise==2.3
protobuf==3.15.5
psutil==5.8.0
py==1.10.0
py-cpuinfo==7.0.0
pycosat==0.6.3
pycparser==2.20
pygame==2.0.1
pyglet==1.5.0
pymongo==3.11.4
pyparsing
PySC2==3.0.0
pytest==6.2.2
python-dateutil
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s2clientprotocol==5.0.6.83830.0
s2protocol==5.0.6.83830.0
sacred==0.8.2
scipy==1.6.1
sentry-sdk==1.0.0
setproctitle==1.2.2
shortuuid==1.0.1
sip==4.19.13
six
sk-video==1.1.10
smmap==3.0.5
snakeviz==2.1.0
subprocess32==3.5.4
tensorboard-logger==0.1.0
tensorboardX==2.2
toml==0.10.2
toolz==0.11.1
torch==1.9.0.dev20210305
torchaudio==0.9.0a0+53fa940
torchvision==0.9.0.dev20210309
tornado
tqdm==4.61.2
typing-extensions
urllib3==1.26.3
wandb==0.10.26
websocket-client==0.58.0
whichcraft==0.6.1
wrapt==1.12.1
